+++ src/domain/career/entity.rs
@@ -2,7 +2,7 @@ use chrono::{NaiveDate};
 
 #[derive(Clone)]
 pub struct CareerEntity {
-  pub user_id: i32,
+  pub user_id: i64,
   pub company: String,
   pub job: String,
   pub in_at: NaiveDate,
@@ -10,7 +10,7 @@ pub struct CareerEntity {
 }
 
 impl CareerEntity {
-  pub fn new(user_id: i32, company: String, job: String, in_at: NaiveDate, out_at: Option<NaiveDate>) -> Self {
+  pub fn new(user_id: i64, company: String, job: String, in_at: NaiveDate, out_at: Option<NaiveDate>) -> Self {
     Self {
       user_id,
       company,
